fix bug

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -1,16 +1,16 @@
 package org.yearup.controllers;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.CategoryDao;
 import org.yearup.data.ProductDao;
 import org.yearup.models.Category;
 import org.yearup.models.Product;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 // add the annotations to make this a REST controller
 // add the annotation to make this controller the endpoint for the following url
@@ -26,6 +26,7 @@ public class CategoriesController
 
 
     // create an Autowired controller to inject the categoryDao and ProductDao
+    @Autowired
     public CategoriesController(CategoryDao categoryDao, ProductDao productDao)
     {
         this.categoryDao = categoryDao;
@@ -37,7 +38,6 @@ public class CategoriesController
     @GetMapping
     public List<Category> getAll()
     {
-        // find and return all categories
         return categoryDao.getAllCategories() ;
     }
 
@@ -45,8 +45,13 @@ public class CategoriesController
     @GetMapping("/{id}")
     public Category getById(@PathVariable int id)
     {
-        // get the category by id
-        return categoryDao.getById(id);
+        Category category = categoryDao.getById(id);
+
+        if (category == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found");
+        }
+
+        return category;
     }
 
     // the url to return all products in category 1 would look like this
@@ -54,50 +59,55 @@ public class CategoriesController
     @GetMapping("/{categoryId}/products")
     public List<Product> getProductsById(@PathVariable int categoryId)
     {
-        // get a list of product by categoryId
-        return productDao.search(categoryId, null, null, null);
+        return productDao.listByCategoryId(categoryId);
     }
 
     // add annotation to call this method for a POST action
     // add annotation to ensure that only an ADMIN can call this function
     @PostMapping
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.CREATED)
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
     public Category addCategory(@RequestBody Category category)
     {
         // insert the category
-        return categoryDao.create(category);
-    }
-    @PostMapping("/public")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Category addCategoryPublic(@RequestBody Category category)
-    {
-        return categoryDao.create(category);
+        try {
+            return categoryDao.create(category);
+        } catch (Exception ex) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 
     // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
     // add annotation to ensure that only an ADMIN can call this function
     @PutMapping("/{id}")
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public Category updateCategory(@PathVariable int id, @RequestBody Category category)
-    {
-        // update the category by id
-        category.setCategoryId(id);
-        categoryDao.update(id, category);
-        return categoryDao.getById(id);
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
+    public void updateCategory(@PathVariable int id, @RequestBody Category category) {
+        try {
+            categoryDao.update(id, category);
+        } catch (Exception ex) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 
 
     // add annotation to call this method for a DELETE action - the url path must include the categoryId
     // add annotation to ensure that only an ADMIN can call this function
     @DeleteMapping("/{id}")
-    //@ResponseStatus(HttpStatus.NO_CONTENT)
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public Map<String, String> deleteCategory(@PathVariable int id) {
-        categoryDao.delete(id);
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
+    @ResponseStatus(HttpStatus.FORBIDDEN)
+    public void deleteCategory(@PathVariable int id)
+    {
+        try {
+            Category category = categoryDao.getById(id);
 
-        Map<String, String> response = new HashMap<>();
-        response.put("message", "Category deleted successfully");
-        return response;
+            if (category == null) {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found");
+            }
+
+            categoryDao.delete(id);
+
+        } catch (Exception ex) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 }
